requirements.txt
flask==2.2.2
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-migrate==4.0.4
flask-sqlalchemy==3.0.2
boto3==1.28.20
pymysql==1.0.3
gunicorn==20.1.0
passlib==1.7.4
python-dotenv==1.0.0
pyyaml==6.0
pytest==7.4.0
Werkzeug==2.3.7
prometheus-client==0.18.0